LayoutTests/imported/w3c/web-platform-tests/css/selectors/focus-visible-013-expected.txt

This test checks that if the active element matches ':focus-visible' and a mouse click causes focus to move elsewhere, the newly focused element should not match :focus-visible.

If the user-agent does not claim to support the :focus-visible pseudo-class then SKIP this test.
Press the TAB key.
If the element that says "Initial" has a red outline, then the test rsult is a FAILURE.
Use the mouse to focus the element below that says "Target".
If the element that says "Target" has a green outline, then the test result is a SUCCESS.
Initial
Target

Harness Error (TIMEOUT), message = null

FAIL :focus-visible does not match after mouse click even if previous focused element was matching :focus-visible assert_equals: outlineColor for DIV#initial should be green expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
